backend/requirements.txt
asgiref==3.8.1
Django==4.2.2
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
pillow==10.3.0
psycopg2-binary==2.9.9
six==1.16.0
sqlparse==0.5.0
transitions==0.9.1
